Added text.py. Updated to allow for button updating.
This modification allows for text created with with the kinetic text tags to be updated along with a button. However, I have not thoroughly tested my modification. While I don't see how it will cause side effects, I can't be certain of that. I'll include more notes on this in the README soon.