Added StyleResearch.txt
These are just my notes on how styles work within the Text class. Attempting to follow the calls from when the style is applied to the text and the path it takes to get there. Included for those interested in learning more about how styles are applied to text and if they would like to find other possible solutions to the TextButton problem.